Add oauth scope information.
We initially thought of leaving this out, allowing for some
flexibility on the provider implmenter's side. But in actually
trying to implement this I found it is a real pain to having to
decide on all this things.

Also, implementation in the instance's software must match the
provider's exactly. So it is really, really helpful to
specify scopes up front.

This also means that in the coming specifications we need to
take great care to think about what is really needed here
carefully.