Replace OAuth 2.0 with HTTP Message Signatures (#27)
* Replace OAuth 2.0 with a very simple bearer token.

This is just one option. While it could be said that this is a
less capable and less secure approach than OAuth, it is less
confusing, easier to implement and comes without the risk
of regular race conditions.

* Custom authentication scheme

This uses an HMAC to authenticate a timestamp. This means that the
secret keys never have to be transmitted after the initial exchange and
that replay attacks can be mitigated.

* Authentication based on HTTP message signatures

Very rough first draft to illustrate how this could work.

* Further flesh out idea.

* Add `Content-Digest` HTTP header and add to signed headers
* Switch to Ed25519 for asymmetric signing
* Completely revamp the registration process to make it easier
  and allow for exchanging public keys.

* Update general/v0.1/protocol_basics.md

Co-authored-by: Renaud Chaput <renchap@gmail.com>

* Add feedback

* Explain "base URL" more and make it discoverable via `nodeinfo`
* Add `Content-Digest` header to example
* Allow for different variations of `redirectUri`

* Incorporate more review feedback.

Clarify a couple of sections and rename client identifiers
to more clearly indicate who they belong to.

* Update OpenAPI spec.

And fix some formatting.

---------

Co-authored-by: Renaud Chaput <renchap@gmail.com>