Clarifications and some simplification
Two unrelated changes here:

1. Clarify what "public" content is as some fediverse software,
e.g. mastodon, has different "levels" of public visibility.

2. Change how backfill requests work. This is informed by an
attempt to actually implement the old version of this. This
version is a bit simpler and a lot more straight-forward to
implement. It gets rid of the "cursor" and only used the
original request's id to get more content.

Conceptually this means there is only one backfill request per
type of content independent of the number of records and FASP
request a continuation of such a request instead of posting many
small ones.

This makes it easier for both parties to keep track of the state
of a request and offers more flexibility for fediverse servers
(i.e. they might still keep track of a "cursor" internally, but
they could also record exactly what was sent previously if they
wish).