Discovery basics: Data Sharing (#36)
* WIP initial draft of discovery fasps.

* Address feedback and add object retrieval section

Most important changes:

* No longer allow to subscribe to individual lifecycle events
* Historical content gets delivered asynchronously using the
  same endpoint as events
* Reword some things to better accommodate for these changes
* Add a section on the actual retrieval of objects

Open questions:

* When fulfilling backfill requests how does the fediverse
  server signal that no more content is available?
* Can we get away with the FASP having no AP inbox/outbox?

* Remove references to AS vocabulary...

... and specify a `moreObjectsAvailable` key to signal when no more
objects are available for a backfill request.

* Change structure

We decided to move to seperate specifications that can be
evolved independently.

Also adjusts to latest requirements from the `general` specification.

* Mention `discoverable` flag.

* Review feedback

* Apply suggestions from code review

Co-authored-by: Emelia Smith <ThisIsMissEm@users.noreply.github.com>

* Fix example

* Add clarifications and new paragraph...

...describing the need to periodically recheck objects.

* Make backfill request consistent with subscription.

* Add missing response

* Add inbox, outbox and webfinger requirements

* Clarifications and some simplification

Two unrelated changes here:

1. Clarify what "public" content is as some fediverse software,
e.g. mastodon, has different "levels" of public visibility.

2. Change how backfill requests work. This is informed by an
attempt to actually implement the old version of this. This
version is a bit simpler and a lot more straight-forward to
implement. It gets rid of the "cursor" and only used the
original request's id to get more content.

Conceptually this means there is only one backfill request per
type of content independent of the number of records and FASP
request a continuation of such a request instead of posting many
small ones.

This makes it easier for both parties to keep track of the state
of a request and offers more flexibility for fediverse servers
(i.e. they might still keep track of a "cursor" internally, but
they could also record exactly what was sent previously if they
wish).

---------

Co-authored-by: Emelia Smith <ThisIsMissEm@users.noreply.github.com>